Minor code fixes

+++ Data/Data/ReleaseDataAccessLayer.cs
@@ -28,10 +28,10 @@ namespace Detention_facility.Data
                 command.Parameters.Add(Constants.ReleaseDate, SqlDbType.DateTime);
                 command.Parameters[Constants.ReleaseDate].Value = release.ReleaseDate;
                 
-                command.Parameters.Add(Constants.AmountPaid, SqlDbType.Int);
+                command.Parameters.Add(Constants.AmountPaid, SqlDbType.Float);
                 command.Parameters[Constants.AmountPaid].Value = release.AmountPaid;
 
-                command.Parameters.Add(Constants.AmountAccrued, SqlDbType.Int);
+                command.Parameters.Add(Constants.AmountAccrued, SqlDbType.Float);
                 command.Parameters[Constants.AmountAccrued].Value = release.AmountAccrued;
 
                 connection.Open();
@@ -62,10 +62,10 @@ namespace Detention_facility.Data
                 command.Parameters.Add(Constants.ReleaseDate, SqlDbType.DateTime);
                 command.Parameters[Constants.ReleaseDate].Value = release.ReleaseDate;
                 
-                command.Parameters.Add(Constants.AmountPaid, SqlDbType.Int);
+                command.Parameters.Add(Constants.AmountPaid, SqlDbType.Float);
                 command.Parameters[Constants.AmountPaid].Value = release.AmountPaid;
 
-                command.Parameters.Add(Constants.AmountAccrued, SqlDbType.Int);
+                command.Parameters.Add(Constants.AmountAccrued, SqlDbType.Float);
                 command.Parameters[Constants.AmountAccrued].Value = release.AmountAccrued;
 
                 connection.Open();
@@ -194,9 +194,9 @@ namespace Detention_facility.Data
 
                     release.ReleaseDate = Convert.ToDateTime(reader.GetValue(1));
 
-                    release.AmountAccrued = reader.GetValue(2) == DBNull.Value ? 0: Convert.ToInt32(reader.GetValue(2)) ;
+                    release.AmountAccrued = reader.GetValue(2) == DBNull.Value ? 0: Convert.ToDouble(reader.GetValue(2)) ;
 
-                    release.AmountPaid = reader.GetValue(3) == DBNull.Value ? 0: Convert.ToInt32(reader.GetValue(3)) ;
+                    release.AmountPaid = reader.GetValue(3) == DBNull.Value ? 0: Convert.ToDouble(reader.GetValue(3)) ;
 
                     release.ReleasedByEmployeeID = Convert.ToInt32(reader.GetValue(4));
 
@@ -238,9 +238,9 @@ namespace Detention_facility.Data
 
                         ReleaseDate = Convert.ToDateTime(reader.GetValue(4)),
 
-                        AmountPaid = Convert.ToInt32(reader.GetValue(5)),
+                        AmountPaid = Convert.ToDouble(reader.GetValue(5)),
 
-                        AmountAccrued = Convert.ToInt32(reader.GetValue(6))
+                        AmountAccrued = Convert.ToDouble(reader.GetValue(6))
                     };
 
                     releases_list.Add(release);
@@ -275,9 +275,9 @@ namespace Detention_facility.Data
 
                     ReleaseDate = Convert.ToDateTime(reader.GetValue(1)),
 
-                    AmountAccrued = reader.GetValue(2) == DBNull.Value ? 0 : Convert.ToInt32(reader.GetValue(2)),
+                    AmountAccrued = reader.GetValue(2) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetValue(2)),
 
-                    AmountPaid = reader.GetValue(3) == DBNull.Value ? 0 : Convert.ToInt32(reader.GetValue(3)),
+                    AmountPaid = reader.GetValue(3) == DBNull.Value ? 0 : Convert.ToDouble(reader.GetValue(3)),
 
                     EmployeeFullName = reader.GetValue(4).ToString(),
                 };
