+++ Models/Models/SmartRelease.cs
@@ -9,9 +9,9 @@ namespace Detention_facility.Models
         [Required]
         public DateTime ReleaseDate { get; set; }
         [Required]
-        public int AmountAccrued { get; set; }
+        public double AmountAccrued { get; set; }
         [Required]
-        public int AmountPaid { get; set; }
+        public double AmountPaid { get; set; }
         [Required]
         public string EmployeeFullName { get; set; }
